+++ NEWS.md
@@ -3,6 +3,7 @@ stevetemplates 0.5.0
 
 The newest release has the following additions and changes to pre-existing templates. Additions include:
 
+- A template for HTML articles. This will allow the conversion of an academic manuscript to a presentable HTML article.
 - A template for Microsoft Word manuscripts. Functionality for Word templates is somewhat limited so I recommend using this only for presentations that need to be anonymized for submission to an academic journal.
 
 Changes include: